pyproject.toml
[project]
name = "capstone-project-team-5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ronit", email = "butironit@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "easygui>=0.98.3",
    "pypdf>=6.1.3",
    "python-docx>=1.2.0",
    "google-genai>=1.46.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.20",
    "openai>=2.8.1",
    "tree-sitter>=0.25.2",
    "tree-sitter-java>=0.23.5",
]

[project.scripts]
capstone-project-team-5 = "capstone_project_team_5:main"

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "ruff>=0.14.0",
]

[tool.ruff]
line-length = 100
indent-width = 4
exclude = [".git", ".ruff_cache", ".venv", ".vscode"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "B",
    "I",
    "UP",
    "SIM",
]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"
docstring-code-format = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra -q"
